requirements.txt
celery>5.3.4
Django>=4.2
djangorestframework>=3.14.0
psycopg2-binary>=2.9.6
djangorestframework-simplejwt>=5.2.2
redis>=5.0.3